# Implement Faster IPv4 Address Retrieval Method