--- a/winfetch.ps1
+++ b/winfetch.ps1
@@ -1380,16 +1380,16 @@ function info_weather {
 function info_local_ip {
     try {
         # Get all network adapters
-        foreach ($ni in [System.Net.NetworkInformation.NetworkInterface]::GetAllNetworkInterfaces()){
+        foreach ($ni in [System.Net.NetworkInformation.NetworkInterface]::GetAllNetworkInterfaces()) {
             # Get the IP information of each adapter
             $properties = $ni.GetIPProperties()
             # Check if the adapter is online, has a gateway address, and the adapter does not have a loopback address
-            if($ni.OperationalStatus -eq 'Up' -and !($null -eq $properties.GatewayAddresses[0]) -and !$properties.GatewayAddresses[0].Address.ToString().Equals("0.0.0.0")){
+            if ($ni.OperationalStatus -eq 'Up' -and !($null -eq $properties.GatewayAddresses[0]) -and !$properties.GatewayAddresses[0].Address.ToString().Equals("0.0.0.0")) {
                 # Check if adapter is a WiFi or Ethernet adapter
-                if ($ni.NetworkInterfaceType -eq "Wireless80211" -or $ni.NetworkInterfaceType -eq "Ethernet"){
-                    foreach ($ip in $properties.UnicastAddresses){
-                        if ($ip.Address.AddressFamily -eq "InterNetwork"){
-                            if(!$local_ip){$local_ip = $ip.Address.ToString()}
+                if ($ni.NetworkInterfaceType -eq "Wireless80211" -or $ni.NetworkInterfaceType -eq "Ethernet") {
+                    foreach ($ip in $properties.UnicastAddresses) {
+                        if ($ip.Address.AddressFamily -eq "InterNetwork") {
+                            if (!$local_ip) { $local_ip = $ip.Address.ToString() }
                         }
                     }
                 }